Adds app/controllers/slack
