+++ lib/slack.js
@@ -14,8 +14,8 @@ function getCallbackIdForCampaign(environmentName, campaign) {
 function getMobileCommonsGroupsText(campaign) {
   const uri = 'https://secure.mcommons.com/groups';
 
-  let text = `Doing: ${uri}/${campaign.mobilecommons_group_doing}`;
-  text = `${text}\nCompleted: ${uri}/${campaign.mobilecommons_group_completed}\n`;
+  let text = `- Doing: ${uri}/${campaign.mobilecommons_group_doing}`;
+  text = `${text}\n- Completed: ${uri}/${campaign.mobilecommons_group_completed}\n`;
 
   return text;
 }
